authorization + get img/video/status from uploader

+++ src/Z_Controller/TrafficStatusInfoController.py
@@ -1,5 +1,7 @@
 from flask import Blueprint, request, jsonify
 from Z_Services.TrafficStatusInfoServices import *
+from Z_Services.UserServices import checkAdmin, checkToken
+from Z_Services.DataServices import findDataByStatusInfoID, findDataByUploaderID
 from pymongo.errors import PyMongoError
 trafficStatusInfo_blueprint = Blueprint('trafficStatusInfo',__name__)
 
@@ -7,21 +9,61 @@ trafficStatusInfo_blueprint = Blueprint('trafficStatusInfo',__name__)
 
 @trafficStatusInfo_blueprint.before_request
 def trafficStatusInfoBeforeRequest():
-    print("before trafficStatusInfo")
+    #Check Access Token
+    access_token = request.headers.get('Authorization')
+    if not access_token:
+        return 'No access token in header', 401
+    try:
+        checkToken(access_token)
+    except Exception as e:
+        print(e)
+        return str(e), 401
+    
 # "TrafficStatusID", "velocity"
 @trafficStatusInfo_blueprint.get('/')
 def getAllTrafficStatusInfo():
     try:
-        res = findAllTrafficStatusInfo()
-        return res
+        access_token = request.headers.get('Authorization')
+        if checkAdmin(access_token):
+            res = findAllTrafficStatusInfo()
+            return res
+        else:
+            return 'Forbidden', 403 
     except Exception as e:
         print(e)
         return str(e), 500
 @trafficStatusInfo_blueprint.get('/<id>')
 def getTrafficStatusInfoID(id):
     try:
-        res = findTrafficStatusInfoByID(id)
-        return res
+        #Tìm uploader:
+        uploader = findDataByStatusInfoID(id)[0]['uploaderID']
+        #Check Token xem có đúng uploader hay admin hay không
+        access_token = request.headers.get('Authorization')
+        if checkAdmin(access_token) or checkToken(access_token)[0] == uploader:
+            res = findTrafficStatusInfoByID(id)
+            return res
+    except Exception as e:
+        print(e)
+        return str(e), 500
+    
+
+@trafficStatusInfo_blueprint.get('/uploader/<id>') #New, chưa thêm swagger
+def getStatusInfoByUploaderID(id):
+    try:
+        #Check Token xem có đúng uploader hay admin hay không
+        access_token = request.headers.get('Authorization')
+        if checkAdmin(access_token) or checkToken(access_token)[0] == id:
+            #Services chỉ có Data theo uploader mà thôi. Trong User có StatusInfoID có thể dựa vào mà tìm.
+            dataList = findDataByUploaderID(id)
+            print(dataList)
+            res = []
+            for data in dataList[0]:
+                statusInfo = findTrafficStatusInfoByID(data['InfoID'])[0]
+                if statusInfo != {}:
+                    res.append(statusInfo)
+            return res
+        else:
+            return 'Forbidden', 403
     except Exception as e:
         print(e)
         return str(e), 500
@@ -37,7 +79,7 @@ def insertTrafficStatusInfoInstance():
             return jsonify({"error": "Bad Request"}), 400
         
         #Trường Required
-        if 'TrafficStatusID' not in trafficStatusInfo or 'velocity' not in trafficStatusInfo:   
+        if 'velocity' not in trafficStatusInfo or 'statuses' not in trafficStatusInfo:   
             return jsonify({"error": "Missing Required Values"}), 400 
         
         print(len(trafficStatusInfo["_id"]))
@@ -46,7 +88,7 @@ def insertTrafficStatusInfoInstance():
 
         #Đảm bảo các trường có đúng không
         for key in trafficStatusInfo.keys():
-            if key not in ["TrafficStatusID", "velocity"]:
+            if key not in ["statuses", "velocity"]:
                 return jsonify({"error": "Wrong key provided"}), 400 
         
         res = insertTrafficStatusInfo(trafficStatusInfo)
@@ -67,8 +109,12 @@ def changeTrafficStatusInfoInstance():
         
         #Đảm bảo các trường có đúng không
         for key in trafficStatusInfo.keys():
-            if key not in ["AccidentFlag", "TrafficJamFlag", "PoliceFlag", "Flooded"]:
+            if key not in ["statuses", "velocity"]:
                 return jsonify({"error": "Wrong key provided"}), 400 
+
+        compareTrafficStatusInfo = findTrafficStatusInfoByID(trafficStatusInfo['_id'])
+        if compareTrafficStatusInfo['TrafficStatusID']!=trafficStatusInfo['TrafficStatusID']:
+            return jsonify({"error": "TrafficStatusID is different from the original one"}), 400
 
         res = updateTrafficStatusInfo(trafficStatusInfo)
         return res
@@ -79,11 +125,16 @@ def changeTrafficStatusInfoInstance():
 @trafficStatusInfo_blueprint.delete('/<id>')
 def deleteTrafficStatusInfoID(id):
     try:
-        print(len(id))
-        if len(id)!=24:
-            return jsonify({"error": "Bad Request"}), 400
-        res = deleteTrafficStatusInfo(id)
-        return res
+        uploader = findDataByStatusInfoID(id)[0]['uploaderID']
+        #Check Token xem có đúng uploader hay admin hay không
+        access_token = request.headers.get('Authorization')
+        if checkAdmin(access_token) or checkToken(access_token)[0] == uploader:
+            if len(id)!=24:
+                return jsonify({"error": "Bad Request"}), 400
+            res = deleteTrafficStatusInfo(id)
+            return res
+        else:
+            return 'Forbidden', 403
     except Exception as e:
         print(e)
         return str(e), 500
